+++ capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require_relative 'NameableDecorator'
+
+class CapitalizeDecorator < NameableDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
