+++ person.rb
@@ -1,33 +1,4 @@
-class Nameable
-  def correct_name
-    raise NotImplementedError, 'Subclasses must implement the correct_name method'
-  end
-end
-
-class NameableDecorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < NameableDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < NameableDecorator
-  def correct_name
-    trimmed_name = @nameable.correct_name[0, 10]
-    trimmed_name.sub!(/\s\w*\z/, '')
-    trimmed_name
-  end
-end
+require_relative 'nameable'
 
 class Person < Nameable
   attr_accessor :name, :age
